Consolidate all methods related to actually running MemMapReduce() on tinymr._base._MRInternal().  Public API methods have all been moved to tinymr.memory.MemMapReduce().
